[IMP] honduras_tuition_plan: Add smaller paper size for collection stub
