+++ honduras_tuition_plan/reports/__init__.py
@@ -1,2 +1,2 @@
 # -*- coding: utf-8 -*-
-from . import res_partner_report_collection_stub
+from . import res_partner_report_collectionstub
